+++ CHANGELOG.md
@@ -17,6 +17,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Changed
 
+- Enable flang architecture flags
+
 ### Deprecated
 
 ## [3.74.0] - 2026-03-09